a script for setting up an el6 rebuild project's bugzilla
The bugzilla-basic-setup-el6 is a manifest of how we set up our el6
bugzilla.  It needs a fair amount of further work before it could
actually be run as a script, but it is probably 90% there, at least
content-wise.  Also, bringing the dag key into the tree that was
used to verify the perl-PatchReader package that was cherrypicked
from repoforge (other than that, just *el6 and epel6 repos required)