checkpoint xbuild progress, i.e. installer iso generator
This commit consists of a couple new directories for general tools
and 4 new files that do not yet achieve the next milestone task
of a pushbutton reproducible *el installable iso generator.

el-build is the main script.  It mirrors the needed input repotrees,
then serves it locally with the user-httpd script, to a throwaway VM it
creates using the min-el60-x64.ks.  The VM will do the pungi processing,
under mock, to generate the ultimate output, a mirrorable release
repotree with installable iso images.