new tools to facilitate the importing of src.rpm deltas into the tree
documentation forthcoming, as well as many many cleanups to this quick
and dirty first pass.  The important thing is that after the subsequent
import, sufficient sha256 checksums will be in place along with these
scripts responsible for the import, to be able to sufficiently audit
things in the future.  Which given the rather small size of the deltas
about to be imported (21MB, 20 of that in the -logos package), isn't
such a big deal anyway (since the delta contents themselves can be
easily audited).

The basic order of things is- run el-build to generate isos, as this
will cause the local tdv srpm repo to be mirrored.  Then run -

sigcheck-tdv-srpms-el6
mirror-tuv-srpms-el6
sigcheck-tuv-srpms-el6
src-pkg-repo-diff--gen-min-srpm-list
src-pkg-repo-diff--gen-debug-output
gen-tdv-srpm-repo

then look in ~/.el-build/cache and enjoy, with the .repo-diff
being what is about to be imported into Ascendos/packages.