f14 key is needed for current mingw32 bootstrapping packages
why these are f14 packages instead of f13 or f15 I know not.  Presumably
the long term mingw32 solution will be a minimal bootstrapping procedure
as per the recent sl-devel discussion with link to fedora packager
providing patches and instructions to centos.  And using f13 to minimize
the sprawl of bootstrapping dependencies (unless f14 really was needed
for some obscure reason)